+++ react/components/ProductSummaryImage/ProductImage.js
@@ -5,7 +5,6 @@ import { CollectionBadges, DiscountBadge } from 'vtex.store-components'
 import classNames from 'classnames'
 
 import ProductSummaryContext from '../ProductSummaryContext'
-import ImageLoader from './ImageLoader'
 import { productShape } from '../../utils/propTypes'
 
 import productSummary from '../../productSummary.css'
@@ -32,6 +31,39 @@ const maybeCollection = ({ productClusters }) => shouldShow => component => {
   }
   return component
 }
+
+export const ImagePlaceholder = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 512 512"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="512" height="512" fill="#F2F2F2" />
+    <rect
+      x="183.857"
+      y="180.2"
+      width="144.286"
+      height="150.474"
+      stroke="#CACBCC"
+      stroke-width="2"
+    />
+    <path d="M183.78 303.688H328.214" stroke="#CACBCC" stroke-width="2" />
+    <path
+      d="M205.082 279.563L223.599 240.507L242.116 260.035L269.892 220.979L306.926 279.563H205.082Z"
+      stroke="#CACBCC"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    />
+    <path
+      d="M252.225 213.939C252.225 219.822 247.66 224.52 242.114 224.52C236.569 224.52 232.004 219.822 232.004 213.939C232.004 208.057 236.569 203.359 242.114 203.359C247.66 203.359 252.225 208.057 252.225 213.939Z"
+      stroke="#CACBCC"
+      stroke-width="2"
+    />
+  </svg>
+)
 
 const ProductImageContent = ({
   product,
@@ -65,7 +97,9 @@ const ProductImageContent = ({
     label: badgeText,
   })
   const withCollection = maybeCollection({ productClusters })
-  const img = <img className={imageContentClassName} src={imageUrl} alt={name} />
+  const img = (
+    <img className={imageContentClassName} src={imageUrl} alt={name} />
+  )
 
   return compose(
     withBadge(showBadge),
@@ -73,7 +107,7 @@ const ProductImageContent = ({
   )(img)
 }
 
-const ProductImage : FunctionComponent = (props) => {
+const ProductImage: FunctionComponent = props => {
   const { product } = useContext(ProductSummaryContext)
   const imageClassName = classNames(productSummary.imageContainer, {
     'db w-100 center': props.displayMode !== 'inline',
@@ -83,7 +117,7 @@ const ProductImage : FunctionComponent = (props) => {
       {path(['sku', 'image', 'imageUrl'], product) ? (
         <ProductImageContent {...props} product={product} />
       ) : (
-        <ImageLoader />
+        <ImagePlaceholder />
       )}
     </div>
   )
@@ -103,7 +137,7 @@ ProductImage.propTypes = {
 ProductImage.defaultProps = {
   showBadge: true,
   showCollections: false,
-  displayMode: 'normal'
+  displayMode: 'normal',
 }
 
 ProductImage.getSchema = () => {