+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Do not render `ProductPrice` when selling price is zero.
+- Remove content loader from `ProductImage` and put a placeholder when the product has no image.
 
 ## [2.19.0] - 2019-04-29
 ### Fixed